requirements.txt
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
pydantic>=2.7.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0
httpx>=0.27.0
tenacity>=8.3.0
pytest>=8.0.0
